src/data/timetable: init json formatted timetable data
Still in early conception, but this is how I hope to structure the data
in the json file to be read. I hope to cache it as well, to store it in
memory until it has been edited, to minimize disk reads and writes.
Maybe using an sql database may be better, but who knows :shrug